managed to add multiple cars in one step, show vehicle shapes
