+++ .gitignore
@@ -6,3 +6,6 @@ __pycache__/sumo_controller.cpython-313.pyc
 __pycache__/detector.cpython-313.pyc
 __pycache__/networking.cpython-313.pyc
 __pycache__/sumo_controller.cpython-313.pyc
+__pycache__/sumo_controller.cpython-312.pyc
+__pycache__/detector.cpython-312.pyc
+__pycache__/networking.cpython-312.pyc
