+++ README.md
@@ -1,4 +1,4 @@
-Tested in an anaconda environment on Windows 10 with the packages listed in the requirements.txt file.
+Tested in an anaconda environment on Windows 11 with the packages listed in the requirements.txt file.
 
 Run each node with the following command format: python main.py -i (insert number 0-3 for node index, required)
                                                                 -v (path to video file, samples provided)
