+++ src/simulation/config.sumocfg
@@ -6,9 +6,10 @@
 <sumoConfiguration xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="http://sumo.dlr.de/xsd/sumoConfiguration.xsd">
 
     <input>
-        <net-file value="test_network.net.xml"/>
-        <route-files value="test_config.rou.xml"/>
+        <net-file value="network.net.xml"/>
+        <route-files value="config.rou.xml"/>
         <additional-files value="sumo_config.xml"/>
+        <gui-settings-file value="settings.xml"/>
     </input>
 
 </sumoConfiguration>
